introduce metaBuilder abstraction
